Dendritic repair
This is a port of BlueRepairSDK dendritic repair code.
It was subject to modifications to better repair data from
/gpfs/bbp.cscs.ch/project/proj55/iavarone/releases/2017-05-15/morphology_release_bk2/05_RepairUnravel/unravel_output/

The tortuosity parameter was refined (old repair was creating dendrites not
tortuous enough)

It adds a soma repulsive factor when sampling the direction of the segment
to better respect the angle distribution between a segment and the
direction going from the center of the soma and the first point of a given
segment.

When continuing an existing section, it still uses a constant diameter but
its value has been changed. It used to be the values of the last point of the
dendrite, now it is taken as the average of the diameters of the last section.
My guess is that the diameter of the last point of cut section is systematically
smaller than what it should be so it does not make sense to take it.

Change-Id: I0b55838268a9c26070e69aa6753b2059f1e768ce