Preparation for oblique repair
The global change is the factorization of the growth code by understanding that
the basal repair and oblique repair only differ by the origin of the sholl
sphere and the branching order offset. This commit does *not* enable oblique
repair yet but only prepare it. Additional test will be added

Change-Id: I86a9dc03a24ed41ec79df74a60ff4bdfc319246c