Decouple repair and cut plane
The repair does not actually needs a cut plane but simply the coordinates
of the points to start regrowing from.

The argument cut_plane of the repair() and Repair.__init__() functions
is replaced by a cut_points one which is the list of point coordinates
to start regrowing from.

This change is especially useful in the context of passing to the repair
points which are not part of a real cut plane. It is important in the context
of implementing the legacy cut detection that will come in the next PR:
https://bbpcode.epfl.ch/source/xref/platform/BlueRepairSDK/BlueRepairSDK/src/helper_dendrite.cpp#691

Also:
1) Removes the full() method as it is overly complex and not really useful.
- It is also not very flexible as it requires a specific directory structure.
- It couples unravelling and repair which should actually be seen as 2
unrelated things.
- At the end it was simply calling unravel_all and then repair_all, which
can easily be done directly by the end user

2) Mute morphio warnings that are rightfully emitted during the repair.