Legacy cut point detection (#26)
Implementing the legacy cut point detection.

In this algorithm, cut points do not belong to a plane but to a half-space. An axis must be passed by the user.
The cut points are the leaves which live in the half-space containing the bigger number of leaves.

This implementation is needed in order to try to reproduce exactly the repair process performed by the old C++ BlueRepairSDK software.