No longer check the cut plane probability
When looking for the intact axons.

This was an enhancement with the new cut plane algorithm but
it is not compatible with the legacy cut plane on which we currently
focus. The old way of computing cut points does not allow for
computing the probability than the cut plane really exists.