Adds the ability to repair only certain type of neurites
The repair function has a new "repair_flags" argument which is a dict of
RepairType -> bool that will disactivate the repair for each RepairType where the bool
is False.

This makes no scientific sense (why would you pick what needs to be repaired ?) but this is how it was implemented before.
https://bbpcode.epfl.ch/source/xref/platform/BlueRepairSDK/BlueRepairSDK/src/repair.cpp#488