Fix a issue in the axon grafting:
* Manually set the grafted section first point as the parent end point

During the axon grafting, donor sections must be appended to existing sections.
They first, must be realigned and then translated so that they start exactly
at the end of the parent section.

Because of some floating point imprecision, sometimes the alignment and
translation of the donor section change slightly (epsilon level) the position of
the first point.

This offset creates some unwanted duplicate points when writing the morphology to disk.