allow custom apical points (#56)
Context:

Repair uses morph-tool to detect apical points, but we want to provide others, if needed. The main use case is if we correct it by hand from a first run or morph-tool, we'd rather use this corrected point.

Resolution:

repair accepts an additional optional argument apical_point to pass the coordinates of the apical points. If not passed the old behavior is used, ie. the apical point is automatically discovered.