error_annotation tool (#54)
This provides a function and a cli to detect possible errors in the reconstruction of morphologies, such as multifurcation, dangling neurites, etc...

This functionality is a sinple call of some NeuroM function, originally implemented in the internal morph-service web application. The cli provides the same as the corresponding morph-service tool, as in, it reads a morphology, and saves an annotated morphology along with a summary report of the detected errors.

Overall, this is to provide a consistent an easy to use error detection, instead of custom codes calling NeuroM functions.